backup_lib.sh: implement and use "adjusted epoch" in prune rules
Prune rules that support divisors (retain X backups every Y minutes/hours/days)
sort backups to buckets by converting timestamps to UNIX time (seconds since
epoch) and then integer dividing them by the bucket period (where period is Y
minutes/hours/days).

This works, but has a quirk: the resulting buckets are UTC-based. In other
words, if a backup has been taken in a +0300 timezone, then the "daily" rule
would consider the "day" to begin and end at 03:00. This is counterintuitive.

To deal with this, use "pseudo epoch" values that are adjusted for timezone
instead of the true epoch.