Revert "btrfs_delete.sh: minor: don't delete snapshot directory itself when cleaning up"
This reverts commit b41fa13ee76e9503933fa2ae004b17f46c41b84e.

It's a wrong thing to do, and leaves empty snapshot directories in place
of deleted snapshot trees which are then picked up by btrfs_list.sh as
if we didn't delete them at all (but, of course, with empty content).

I somewhy thought that the root of this search is `/snapshots` itself,
which it makes sense to not delete, but in fact the root of the search
is the snapshot directory itself (the one named after ISO 8601 id).