*.sh: enable lastpipe and use pipelines instead of here-processes
There is a lovely syntax called here-processes, or `<(...)`, which allows us to
handle a pipeline output in the current shell, like this:

We use it to handle a pipeline output in current shell, like this:

    while read r; do
        output+="$r"
    done < <(cat ... | sort ...)

Unfortunately, it suffers from a critical flaw: it is impossible to catch errors
in the subshell (errexit or pipefail do not work).

Thankfully, bash 4.2+ have a shopt called lastpipe which makes bash execute the
_last_ pipeline stage in foreground, so that code like

    cat ... | sort ... | while read r; do
        output+="$r"
    done

will finally do what you naturally expect it to do.

So, get rid of `< <(...)` and use ordinary pipelines with `lastpipe` and
`pipefail`.